services/capture/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=services/capture/requirements.txt --pip-args=None services/capture/requirements.in
#
python-dotenv==1.1.1
    # via -r services/capture/requirements.in
scapy==2.6.1
    # via -r services/capture/requirements.in
